DRMCP - Fix recipe using the gen ai library tool import
