+++ src/datarobot_genai/drmcp/core/routes.py
@@ -21,7 +21,7 @@ from starlette.responses import JSONResponse
 from .dynamic_tools.deployment.controllers import delete_registered_tool_deployment
 from .dynamic_tools.deployment.controllers import get_registered_tool_deployments
 from .dynamic_tools.deployment.controllers import register_tool_for_deployment_id
-from .memory_management import get_memory_manager
+from .memory_management.manager import get_memory_manager
 from .routes_utils import prefix_mount_path
 
 logger = getLogger(__name__)
